requirements.txt
absl-py==0.10.0
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
altair==4.2.2
aniso8601==9.0.1
antlr4-python3-runtime==4.9.3
anyio==3.6.2
appdirs==1.4.4
appnope==0.1.0
argilla==1.6.0
argon2-cffi==20.1.0
asgiref==3.3.1
astor==0.8.1
astroid==2.11.2
asttokens==2.2.1
async-generator==1.10
async-timeout==4.0.2
attrs==20.2.0
Babel==2.9.0
backcall==0.2.0
backoff==2.2.1
base58==2.0.1
bcrypt==4.0.1
beautifulsoup4==4.9.3
bidict==0.22.0
black==21.4b2
bleach==3.2.1
blinker==1.6.2
blis==0.7.9
boto3==1.15.2
botocore==1.18.2
bottle==0.12.19
bs4==0.0.1
CacheControl==0.12.6
cachetools==4.1.1
cachy==0.3.0
catalogue==2.0.8
certifi==2023.5.7
cffi==1.14.3
chardet==3.0.4
charset-normalizer==3.1.0
chromadb==0.3.21
cleo==0.7.6
click==8.1.3
clickhouse-connect==0.5.23
clikit==0.4.3
cloudpickle==2.2.1
colorama==0.4.6
coloredlogs==15.0.1
commonmark==0.9.1
confection==0.0.4
conllu==4.2.1
contourpy==1.0.7
cryptography==40.0.2
cycler==0.10.0
cymem==2.0.3
Cython==0.29.21
dataclasses-json==0.5.7
datasets==1.1.2
decorator==4.4.2
defusedxml==0.6.0
Deprecated==1.2.13
detectron2 @ git+https://github.com/facebookresearch/detectron2.git@d1e04565d3bec8719335b88be9e9b961bf3ec464
dialogflow==1.1.0
dill==0.3.2
distlib==0.3.1
Django==3.1.3
dm-tree==0.1.5
dnspython==2.3.0
docopt==0.6.2
duckdb==0.7.1
effdet==0.3.0
elasticsearch==7.9.1
email-validator==2.0.0.post2
en-core-web-lg==2.3.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl#sha256=0964370218b7e1672a30ac50d72cdc6b16f7c867496f1d60925691188f4d2510
entrypoints==0.3
enum-compat==0.0.3
et-xmlfile==1.1.0
executing==1.2.0
faiss-cpu==1.6.3
fastapi==0.95.1
ffmpy==0.3.0
filelock==3.0.12
filetype==1.2.0
fire==0.3.1
Flask==2.3.2
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
Flask-Login==0.6.2
Flask-RESTful==0.3.9
flask-simplelogin==0.1.2
Flask-SocketIO==5.1.1
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
flatbuffers==23.3.3
fonttools==4.39.4
frozenlist==1.3.3
fsspec==2023.5.0
future==0.18.2
fvcore==0.1.5.post20221221
gcloud==0.18.3
gensim==3.8.3
git-python==1.0.3
gitdb==4.0.5
GitPython==3.1.10
google-api-core==1.28.0
google-auth==1.30.1
google-auth-oauthlib==0.4.1
google-cloud-dialogflow==2.3.0
googleapis-common-protos==1.52.0
gradio==3.28.3
gradio_client==0.2.0
graphviz==0.20.1
greenlet==2.0.2
grpcio==1.33.1
gunicorn==20.1.0
h11==0.14.0
hnswlib==0.7.0
html2text==2020.1.16
html5lib==1.1
httpcore==0.16.3
httplib2==0.19.0
httptools==0.5.0
httpx==0.23.3
huggingface-hub==0.14.1
humanfriendly==10.0
hydra-core==1.3.2
icecream==2.1.3
idna==2.10
ikpy==3.1
imageio==2.28.1
imageio-ffmpeg==0.4.8
importlib-metadata==6.6.0
importlib-resources==5.12.0
iniconfig==1.1.1
iopath==0.1.9
ipykernel==5.3.4
ipython==7.18.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==5.5.2
itsdangerous==2.1.2
jedi==0.17.2
Jinja2==3.1.2
jmespath==0.10.0
joblib==0.16.0
json5==0.9.5
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.6.3
jupyter-server==1.3.0
jupyterlab==3.0.7
jupyterlab-pygments==0.1.2
jupyterlab-server==2.2.0
jwcrypto==0.8
jwt==1.0.0
keras==2.11.0
keyring==20.0.1
kiwisolver==1.2.0
langchain==0.0.161
langcodes==3.3.0
layoutparser==0.3.4
lazy-object-proxy==1.4.3
leveldb==0.201
linkify-it-py==2.0.2
lockfile==0.12.2
lxml==4.9.2
lz4==4.3.2
Markdown==3.3.2
markdown-it-py==2.2.0
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.7.1
mccabe==0.6.1
mdit-py-plugins==0.3.3
mdurl==0.1.2
mistune==0.8.4
monotonic==1.6
more-itertools==8.8.0
moviepy==1.0.3
mpmath==1.2.1
msg-parser==1.2.0
msgpack==1.0.0
multidict==6.0.4
multiprocess==0.70.10
murmurhash==1.0.2
music21==6.7.1
mypy-extensions==0.4.3
nbclassic==0.2.6
nbclient==0.5.0
nbconvert==6.0.7
nbformat==5.0.7
nest-asyncio==1.4.1
networkx==3.1
nlp==0.4.0
nltk==3.5
notebook==6.1.4
numexpr==2.8.4
numpy==1.23.5
oauth2client==4.1.3
oauthlib==3.1.0
olefile==0.46
omegaconf==2.3.0
onnxruntime==1.14.1
openai==0.27.4
openapi-schema-pydantic==1.2.4
opencv-python==4.5.1.48
openpyxl==3.1.2
orjson==3.8.12
packaging==23.1
pandas==2.0.1
pandocfilters==1.4.2
parso==0.7.1
pastel==0.2.1
pathspec==0.9.0
pathtools==0.1.2
pathy==0.10.1
pdf2image==1.16.3
pdfminer.six==20221105
pdfplumber==0.9.0
pexpect==4.8.0
piazza-api==0.11.0
pickleshare==0.7.5
Pillow==9.5.0
pipreqs==0.4.13
pkginfo==1.5.0.1
plac==1.1.3
platformdirs==2.5.1
pluggy==0.13.1
poetry==1.0.10
poppy-creature==2.0.0
portalocker==2.0.0
posthog==3.0.1
praw==7.7.0
prawcore==2.3.0
preshed==3.0.2
prettytable==2.0.0
proglog==0.1.10
prometheus-client==0.8.0
promise==2.3
prompt-toolkit==3.0.7
proto-plus==1.18.1
protobuf==3.13.0
psutil==5.7.2
ptyprocess==0.6.0
py==1.9.0
pyarrow==2.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.6
pycparser==2.20
pycrypto==2.6.1
pydantic==1.10.7
pydeck==0.5.0b1
pydot==1.4.2
pydub==0.25.1
Pygments==2.15.1
PyJWT==2.7.0
pylev==1.3.0
pylint==2.13.5
PyMySQL==1.0.2
pypandoc==1.11
pyparsing==2.4.7
pypot==3.1.3
pyrsistent==0.14.11
pyserial==3.5
pytesseract==0.3.10
pytest==6.1.1
pytextrank==3.2.4
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
python-engineio==4.3.1
python-jwt==3.3.0
python-magic==0.4.27
python-multipart==0.0.6
python-pptx==0.6.21
python-socketio==5.5.2
pytorch-lightning==0.9.0
pytz==2020.1
PyYAML==6.0
pyzmq==19.0.2
qtconsole==4.7.7
QtPy==1.9.0
quickumls==1.4.0.post1
quickumls-simstring==1.1.5.post1
rank-bm25==0.2.1
regex==2023.5.5
requests==2.30.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rfc3986==1.5.0
rich==13.0.1
rouge-score==0.0.4
rsa==4.6
s3transfer==0.3.3
sacrebleu==1.4.14
sacremoses==0.0.43
scikit-learn==0.23.2
scipy==1.10.1
seaborn==0.11.0
semantic-version==2.10.0
Send2Trash==1.5.0
sentence-transformers==2.2.2
sentencepiece==0.1.91
seqeval==1.2.1
shellingham==1.3.2
six==1.15.0
sklearn==0.0
smart-open==6.3.0
smmap==3.0.4
sniffio==1.2.0
soupsieve==2.1
spacy==3.5.2
spacy-legacy==3.0.12
spacy-loggers==1.0.4
SQLAlchemy==2.0.12
sqlparse==0.4.1
srsly==2.4.6
sseclient==0.0.24
starlette==0.26.1
streamlit==0.69.2
sympy==1.7.1
tabulate==0.9.0
tenacity==8.2.2
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow-datasets==4.0.1
tensorflow-metadata==0.24.0
termcolor==1.1.0
terminado==0.9.1
testpath==0.4.4
thinc==8.1.10
threadpoolctl==2.1.0
tiktoken==0.4.0
timm==0.6.13
tokenizers==0.13.3
toml==0.10.1
tomli==2.0.1
tomlkit==0.5.11
toolz==0.11.1
torch==1.12.1
torchvision==0.13.1
tornado==6.1
tqdm==4.49.0
traitlets==5.0.4
transformers==4.28.1
typer==0.7.0
typing-inspect==0.8.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==2.1
uc-micro-py==1.0.2
Unidecode==1.1.1
unqlite==0.8.1
unstructured==0.6.3
unstructured-inference==0.4.4
update-checker==0.18.0
urllib3==2.0.2
uvicorn==0.22.0
uvloop==0.17.0
validators==0.18.1
virtualenv==20.0.35
waitress==2.1.2
Wand==0.6.11
wasabi==1.1.1
watchdog==3.0.0
watchfiles==0.19.0
wcwidth==0.2.5
webcolors==1.11.1
webencodings==0.5.1
websocket-client==1.5.1
websockets==11.0.3
Werkzeug==2.3.4
widgetsnbextension==3.5.1
wikipedia==1.4.0
wrapt==1.14.1
WTForms==3.0.1
XlsxWriter==3.1.0
xxhash==2.0.0
yacs==0.1.8
yarg==0.1.9
yarl==1.8.2
zipp==3.15.0
zstandard==0.21.0
